[win32gdk] Port to Win64
Replace several `long` by `__int64` and/or `intptr_t`, `Long_t` by `Longptr_t`, `int` by `size_t`, and fix pointer formatting

+++ graf2d/win32gdk/gdk/src/gdk/gdkselection.h
@@ -47,7 +47,7 @@ extern "C" {
                                        guchar ** data,
                                        GdkAtom * prop_type,
                                        gint * prop_format);
-   void gdk_selection_send_notify(guint32 requestor,
+   void gdk_selection_send_notify(gulong requestor,
                                   GdkAtom selection,
                                   GdkAtom target,
                                   GdkAtom property, guint32 time);
